entrypoint 위치 app/ 으로 변경
